# apps/web/openapi.json
{
  "components": {
    "schemas": {
      "AbstentionDetail": {
        "additionalProperties": false,
        "description": "Why no claim was rendered, in terms a reader can act on.\n\n``closest_evidence_ids`` names what came closest; ``closest_evidence`` carries the\ncanonical record for each one where it could be resolved. The two are separate because\nthey can legitimately diverge: an abstention raised *because* detail was unavailable\nknows the identifiers and by definition cannot resolve them, and a reader is still\nbetter served by the identifiers than by silence.\n\nNothing here passed a claim gate. The detail exists so a reader can open a near miss\nand judge whether the corpus is thin or the question was wrong - not so it can be read\nas support.",
        "properties": {
          "closest_evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceDetail"
            },
            "maxItems": 20,
            "title": "Closest Evidence",
            "type": "array"
          },
          "closest_evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Closest Evidence Ids",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "missing_evidence_roles": {
            "items": {
              "type": "string"
            },
            "title": "Missing Evidence Roles",
            "type": "array"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          }
        },
        "required": [
          "reason_code",
          "message"
        ],
        "title": "AbstentionDetail",
        "type": "object"
      },
...
      "ActiveCorpusRelease": {
        "additionalProperties": false,
        "properties": {
          "activated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "corpus_release_id": {
            "title": "Corpus Release Id",
            "type": "string"
          },
          "manifest_sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Manifest Sha256",
            "type": "string"
          },
          "qdrant_collection": {
            "title": "Qdrant Collection",
            "type": "string"
          },
          "serving_mode": {
            "$ref": "#/components/schemas/ReleaseServingMode",
            "default": "ACTIVATED"
          }
        },
        "required": [
          "corpus_release_id",
          "manifest_sha256",
          "qdrant_collection"
        ],
        "title": "ActiveCorpusRelease",
        "type": "object"
      },
...
      "GuidelineConflict": {
        "additionalProperties": false,
        "description": "A disagreement between passages, typed rather than stringly.\n\n``evidence_ids`` was previously flattened to a comma-joined string inside an untyped\ndict, which every client had to re-parse and none could rely on. Typed, the two\npassages a conflict is *about* are addressable, which is what lets an interface show\nthem side by side instead of describing them.\n\nEvery field except the identifiers is optional: a record that arrives without a\nrecognised type is still shown, as unclassified, with whatever it does carry. The\nsystem does not resolve the disagreement - it reports it.",
        "properties": {
          "conflict_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflict Type"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Evidence Ids",
            "type": "array"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "title": "GuidelineConflict",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QuestionAccepted": {
        "additionalProperties": false,
        "properties": {
          "question_id": {
            "title": "Question Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionStatus"
          }
        },
        "required": [
          "question_id",
          "status"
        ],
        "title": "QuestionAccepted",
        "type": "object"
      },
      "QuestionContextPatch": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "$ref": "#/components/schemas/ClinicalContext"
          }
        },
        "required": [
          "context"
        ],
        "title": "QuestionContextPatch",
        "type": "object"
      },
      "QuestionCreate": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "question": {
            "maxLength": 4000,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          },
          "source_filters": {
            "$ref": "#/components/schemas/SourceFilters"
          }
        },
        "required": [
          "question"
        ],
        "title": "QuestionCreate",
        "type": "object"
      },
      "QuestionResult": {
        "additionalProperties": false,
        "properties": {
          "abstention": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AbstentionDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/RenderedClaim"
            },
            "maxItems": 100,
            "title": "Claims",
            "type": "array"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/GuidelineConflict"
            },
            "maxItems": 50,
            "title": "Conflicts",
            "type": "array"
          },
          "corpus_release": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActiveCorpusRelease"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence_details": {
            "items": {
              "$ref": "#/components/schemas/EvidenceDetail"
            },
            "maxItems": 100,
            "title": "Evidence Details",
            "type": "array"
          },
          "interpreted_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClinicalContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "question_id": {
            "title": "Question Id",
            "type": "string"
          },
          "retrieval_candidates": {
            "items": {
              "$ref": "#/components/schemas/RetrievalCandidate"
            },
            "maxItems": 100,
            "title": "Retrieval Candidates",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "verification_summary": {
            "$ref": "#/components/schemas/VerificationSummary"
          }
        },
        "required": [
          "question_id",
          "question",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "QuestionResult",
        "type": "object"
      },
      "QuestionStatus": {
        "enum": [
          "QUEUED",
          "CONTEXT_EXTRACTED",
          "RETRIEVING",
          "RERANKING",
          "SEARCHING_COUNTER_EVIDENCE",
          "CHECKING_EVIDENCE_COMPLETENESS",
          "VERIFYING",
          "ANSWER_READY",
          "ABSTAINED",
          "FAILED"
        ],
        "title": "QuestionStatus",
        "type": "string"
      },
...
      "RenderedClaim": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "minLength": 1,
            "title": "Claim Id",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Evidence Ids",
            "type": "array"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "verification_status": {
            "minLength": 1,
            "title": "Verification Status",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "text",
          "evidence_ids",
          "verification_status"
        ],
        "title": "RenderedClaim",
        "type": "object"
      },
      "RetrievalCandidate": {
        "additionalProperties": false,
        "description": "A retrieved passage that no rendered claim cites.\n\nUncited evidence reaches a client through this field and nowhere else. Keeping it\nout of ``claims`` is the point: a candidate carries the rank it held in retrieval,\nso a reader can see how far past the cited support they have gone, and it can never\nbe mistaken for the evidence a claim was verified against. Ranks are sparse by\nconstruction - the cited passages are the gaps.",
        "properties": {
          "evidence_id": {
            "minLength": 1,
            "title": "Evidence Id",
            "type": "string"
          },
          "retrieval_rank": {
            "minimum": 1.0,
            "title": "Retrieval Rank",
            "type": "integer"
          }
        },
        "required": [
          "evidence_id",
          "retrieval_rank"
        ],
        "title": "RetrievalCandidate",
        "type": "object"
      },
...
      "SourceFilters": {
        "additionalProperties": false,
        "properties": {
          "jurisdictions": {
            "items": {
              "type": "string"
            },
            "title": "Jurisdictions",
            "type": "array"
          },
          "organizations": {
            "items": {
              "type": "string"
            },
            "title": "Organizations",
            "type": "array"
          }
        },
        "title": "SourceFilters",
        "type": "object"
      },
...
      "TableRowNeighbour": {
        "additionalProperties": false,
        "description": "One row of a table, positioned relative to the row that was asked about.",
        "properties": {
          "evidence": {
            "$ref": "#/components/schemas/EvidenceDetail"
          },
          "is_anchor_row": {
            "title": "Is Anchor Row",
            "type": "boolean"
          },
          "row_index": {
            "minimum": 0.0,
            "title": "Row Index",
            "type": "integer"
          },
          "row_number": {
            "minimum": 1.0,
            "title": "Row Number",
            "type": "integer"
          }
        },
        "required": [
          "row_index",
          "row_number",
          "is_anchor_row",
          "evidence"
        ],
        "title": "TableRowNeighbour",
        "type": "object"
      },
      "TableRowNeighbourhood": {
        "additionalProperties": false,
        "description": "The rows immediately around an anchored row of one table.\n\nAn empty `rows` is an ordinary answer, not an error: the release may hold no\nneighbours, the table may not be in it, or the source may not be servable. The caller\nis told the same thing in each case, for the reason `sources.py` gives.",
        "properties": {
          "anchor_row_index": {
            "minimum": 0.0,
            "title": "Anchor Row Index",
            "type": "integer"
          },
          "radius": {
            "minimum": 0.0,
            "title": "Radius",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/TableRowNeighbour"
            },
            "title": "Rows",
            "type": "array"
          },
          "source_id": {
            "minLength": 1,
            "title": "Source Id",
            "type": "string"
          },
          "table_id": {
            "minLength": 1,
            "title": "Table Id",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "table_id",
          "anchor_row_index",
          "radius"
        ],
        "title": "TableRowNeighbourhood",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerificationSummary": {
        "additionalProperties": false,
        "properties": {
          "rendered_claims": {
            "default": 0,
            "title": "Rendered Claims",
            "type": "integer"
          },
          "supported_claims": {
            "default": 0,
            "title": "Supported Claims",
            "type": "integer"
          },
          "withheld_claims": {
            "default": 0,
            "title": "Withheld Claims",
            "type": "integer"
          }
        },
        "title": "VerificationSummary",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Ingestion-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "summary": "Research-only evidence-gated guideline QA API",
    "title": "Guideline Evidence QA",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/questions": {
      "post": {
        "operationId": "submit_question_v1_questions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Question",
        "tags": [
          "questions"
        ]
      }
    },
    "/v1/questions/{question_id}": {
      "get": {
        "operationId": "get_question_v1_questions__question_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Question",
        "tags": [
          "questions"
        ]
      }
    },
    "/v1/questions/{question_id}/context": {
      "patch": {
        "operationId": "replace_question_context_v1_questions__question_id__context_patch",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionContextPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Question Context",
        "tags": [
          "questions"
        ]
      }
    },
    "/v1/questions/{question_id}/events": {
      "get": {
        "operationId": "question_events_v1_questions__question_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Question Events",
        "tags": [
          "questions"
        ]
      }
    },
    "/v1/sources/{source_id}/pages/{page_number}": {
      "get": {
        "operationId": "get_source_page_v1_sources__source_id__pages__page_number__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "page_number",
            "required": true,
            "schema": {
              "title": "Page Number",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "dpi",
            "required": false,
            "schema": {
              "default": 144,
              "maximum": 300,
              "minimum": 72,
              "title": "Dpi",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/png": {}
            },
            "description": "Rendered page image"
          },
          "404": {
            "description": "No page image is available for this source"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Source Page",
        "tags": [
          "sources"
        ]
      }
    },
    "/v1/sources/{source_id}/tables/{table_id}/rows": {
      "get": {
        "description": "The rows around a cited row of one table.\n\nA spreadsheet row is this corpus's unit of evidence, and one row read alone is often\nnot decidable - the row above opens the condition, the row below carries the\nexception. This is the neighbourhood a reader needs to check a citation, served from\nthe same release the answer came from.\n\nIt performs no licence act the answer did not already perform: each row is projected\nthrough the same path as cited evidence, so a source that may not be excerpted yields\nneighbours with addresses and no text. Every failure answers with an empty window\nrather than a status, on the same grounds as the page route above: which rows a\nrelease does not contain is a fact about a corpus the caller has not been granted.",
        "operationId": "get_table_row_neighbourhood_v1_sources__source_id__tables__table_id__rows_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          },
          {
            "description": "Zero-based row index to centre the window on",
            "in": "query",
            "name": "row",
            "required": true,
            "schema": {
              "description": "Zero-based row index to centre the window on",
              "minimum": 0,
              "title": "Row",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "radius",
            "required": false,
            "schema": {
              "default": 3,
              "maximum": 10,
              "minimum": 0,
              "title": "Radius",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableRowNeighbourhood"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Table Row Neighbourhood",
        "tags": [
          "sources"
        ]
      }
    }
  }
}